refactor(runtime): split terminal lifecycle and app loop helpers
