+++ docs/architecture.md
@@ -5,7 +5,8 @@ This crate is organized around a small set of layers.
 - `core`: shared, dependency-light model types that multiple layers need.
 - `fs` and `file_info`: filesystem access, file classification, and metadata discovery.
 - `preview`: preview construction and rendering-oriented preview data.
-- `app`: runtime coordination, state, jobs, and user actions.
+- `app`: application state, jobs, and user actions.
+- `runtime`: application runner, terminal lifecycle, event loop, drawing glue, and session output.
 - `ui`: terminal rendering, layout, and theming.
 
 Current boundary rules:
@@ -16,7 +17,7 @@ Current boundary rules:
 - `preview` is presentation code, but it should not depend on `app`.
 - `preview` should not reach into `ui::theme` directly. The explicit adapter boundary for theme
   access is `src/preview/appearance.rs`.
-- `app` coordinates behavior; it should not be the home for generic data model types that other
+- `app` owns behavior; it should not be the home for generic data model types that other
   layers need.
 
 These rules are enforced by the architecture guardrail test and CI. Keep this document focused on
